+++ src/finlang.test/TranspilerTest/C99TranspilerTest.cs
@@ -21,13 +21,13 @@ public class C99TranspilerTest
         transpiler.GatherSolutionDeclarations();
 
         transpiler.Generate();
-        var ledCls = transpiler.c99ClassEnum.Single(c => c.GetFqn() == "hal.Led");
+        var ledCls = transpiler.c99ClassesEnums.Single(c => c.GetFqn() == "hal.Led");
         string ledStructCode = ledCls.hFile.mainCode.ToString();
         ledStructCode.Should().Contain("typedef struct hal_Led hal_Led;");
         ledStructCode.Should().Contain("  hal_Gpio * _gpio;");
-        ledCls.hFile.fqnDependencies.Should().BeEquivalentTo("hal.Gpio");
+        ledCls.hFile.fqnDependencies.Should().BeEquivalentTo("hal.Gpio", "finlang.u8");
 
-        var mainAppCls = transpiler.c99ClassEnum.Single(c => c.GetFqn() == "app.MainApp");
+        var mainAppCls = transpiler.c99ClassesEnums.Single(c => c.GetFqn() == "app.MainApp");
         string mainAppStructCode = mainAppCls.hFile.mainCode.ToString();
         mainAppStructCode.Should().Contain("typedef struct app_MainApp app_MainApp;");
         mainAppStructCode.Should().Contain("  uint16_t period_ms;");
@@ -50,7 +50,8 @@ public class C99TranspilerTest
             "app_Counter.h",
             "app_Counter.c",
             "hal_Helper.h",
-            "hal_Helper.c"
+            "hal_Helper.c",
+            "hal_GpioPinState.h"
         );
     }
 }