+++ src/test/ExSln2/LedBlinker/hal/GpioPinState.cs
@@ -1,0 +1,7 @@
+namespace hal;
+
+public enum GpioPinState
+{
+    Low,
+    High
+}