+++ src/test/ExSln2/c99/gen/app_MainApp.h
@@ -1,4 +1,4 @@
-// finlang generated file for c# app.MainApp class
+// finlang generated file for c# app.MainApp type
 #pragma once
 
 #include <stdint.h>