+++ src/test/ExSln2/c99/gen/hal_Gpio.h
@@ -1,12 +1,12 @@
-// finlang generated file for c# hal.Gpio class
+// finlang generated file for c# hal.Gpio type
 #pragma once
 
 #include "hal_Gpio_port_implementation.h" // You need to provide this
-#include <stdbool.h>
+#include "hal_GpioPinState.h"
 
 
 // Class is a Foreign Function Interface. No struct generated.
 
-bool hal_Gpio_read(hal_Gpio * self);
+hal_GpioPinState hal_Gpio_read(hal_Gpio * self);
 
-void hal_Gpio_write(hal_Gpio * self, bool state);
+void hal_Gpio_write(hal_Gpio * self, hal_GpioPinState state);