+++ src/test/ExSln2/c99/gen/hal_GpioPinState.h
@@ -1,0 +1,11 @@
+// finlang generated file for c# hal.GpioPinState type
+#pragma once
+
+
+
+
+typedef enum hal_GpioPinState
+{
+    hal_GpioPinState_Low,
+    hal_GpioPinState_High
+} hal_GpioPinState;