+++ src/test/ExSln2/c99/gen/hal_Led.c
@@ -1,4 +1,4 @@
-// finlang generated file for c# hal.Led class
+// finlang generated file for c# hal.Led type
 
 #include "hal_Led.h"
 #include <string.h>
@@ -14,12 +14,12 @@ void hal_Led_ctor(hal_Led * self, hal_Gpio * gpio)
 // Will toggle the state of the LED
 void hal_Led_toggle(hal_Led * self)
 {
-    if (hal_Gpio_read(self->_gpio))
+    if (hal_Gpio_read(self->_gpio) == hal_GpioPinState_High)
     {
-        hal_Gpio_write(self->_gpio, false); // Turn off
+        hal_Gpio_write(self->_gpio, hal_GpioPinState_Low); // Turn off
     }
     else
     {
-        hal_Gpio_write(self->_gpio, true); // Turn on
+        hal_Gpio_write(self->_gpio, hal_GpioPinState_High); // Turn on
     }
 }