+++ src/test/ExSln2/c99/hal_Gpio_port_implementation.c
@@ -1,14 +1,14 @@
 #include "hal_Gpio.h"
 #include <stdio.h>
 
-bool fake_gpio_state = false;
+hal_GpioPinState fake_gpio_state = hal_GpioPinState_Low;
 
-bool hal_Gpio_read(hal_Gpio * self)
+hal_GpioPinState hal_Gpio_read(hal_Gpio * self)
 {
     return fake_gpio_state;
 }
 
-void hal_Gpio_write(hal_Gpio * self, bool state)
+void hal_Gpio_write(hal_Gpio * self, hal_GpioPinState state)
 {
     fake_gpio_state = state;
     printf("hal_Gpio_write(%d)\n", state);